requirements.txt
python-dotenv 
django 
djangorestframework
mimesis
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2==2.8.5
pytz
whitenoise==5.1.0
